refactor(skippy-protocol): split protocol surface
