+++ crates/skippy-protocol/src/lib.rs
@@ -1,5 +1,3 @@
-use serde::{Deserialize, Serialize};
-
 /// Compatibility namespace for tokenizer contracts.
 ///
 /// New consumers should depend on `skippy-tokenizer` directly. Keeping this
@@ -16,648 +14,30 @@ pub mod proto {
     }
 }
 
-pub const SCHEMA_VERSION: u32 = 1;
-pub const STAGE_ALPN_V2: &[u8] = b"skippy-stage/2";
-pub const STAGE_SUBPROTOCOL_NAME: &str = "skippy-stage";
-pub const STAGE_SUBPROTOCOL_MAJOR: u32 = 2;
-pub const STAGE_SUBPROTOCOL_FEATURE_STAGE_CONTROL: &str = "stage-control";
-pub const STAGE_PROTOCOL_GENERATION: u32 = 4;
-/// Generation-scoped stage capability. A peer can advertise `stage-control`
-/// while still rejecting current-generation frames, so split planning gates on
-/// this exact token before sending current-generation control requests.
-pub const STAGE_SUBPROTOCOL_FEATURE_STAGE_PROTOCOL_GENERATION_V4: &str = "stage-generation-4";
-pub const STAGE_SUBPROTOCOL_FEATURE_STAGE_GENERATION: &str =
-    STAGE_SUBPROTOCOL_FEATURE_STAGE_PROTOCOL_GENERATION_V4;
-pub const STAGE_SUBPROTOCOL_FEATURE_ARTIFACT_TRANSFER: &str = "artifact-transfer";
-pub const STAGE_SUBPROTOCOL_FEATURE_STATUS_LIST: &str = "status-list";
-pub const STAGE_STREAM_CONTROL: u8 = 0x01;
-pub const STAGE_STREAM_TRANSPORT: u8 = 0x02;
-pub const STAGE_STREAM_ARTIFACT_TRANSFER: u8 = 0x03;
-pub const MAX_STAGE_FRAME_BYTES: usize = 8 * 1024 * 1024;
-/// Maximum number of unresolved verify windows covered by native checkpoints.
-pub const MAX_VERIFY_WINDOW_PIPELINE_DEPTH: usize = 64;
+mod config;
+mod messages;
+mod validation;
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum StageFrameError {
-    BadGeneration { got: u32 },
-    InvalidEndpointId { got: usize },
-    InvalidArtifactDigestLength { got: usize },
-    InvalidArtifactPath,
-    InvalidArtifactOffset,
-    MissingStageControlCommand,
-    MissingStageControlResponse,
-    MissingStageTransportTarget,
-    MissingStageArtifactTarget,
-}
-
-impl std::fmt::Display for StageFrameError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            StageFrameError::BadGeneration { got } => write!(
-                f,
-                "bad skippy stage generation: expected {}, got {}",
-                STAGE_PROTOCOL_GENERATION, got
-            ),
-            StageFrameError::InvalidEndpointId { got } => {
-                write!(f, "invalid endpoint_id length: expected 32, got {got}")
-            }
-            StageFrameError::InvalidArtifactDigestLength { got } => write!(
-                f,
-                "invalid artifact sha256 length: expected 64 hex chars, got {got}"
-            ),
-            StageFrameError::InvalidArtifactPath => {
-                write!(f, "artifact relative_path must be a safe relative path")
-            }
-            StageFrameError::InvalidArtifactOffset => {
-                write!(f, "artifact offset exceeds expected artifact size")
-            }
-            StageFrameError::MissingStageControlCommand => {
-                write!(f, "stage control command is required but missing")
-            }
-            StageFrameError::MissingStageControlResponse => {
-                write!(f, "stage control response is required but missing")
-            }
-            StageFrameError::MissingStageTransportTarget => {
-                write!(f, "stage transport target is required but missing")
-            }
-            StageFrameError::MissingStageArtifactTarget => {
-                write!(f, "stage artifact transfer target is required but missing")
-            }
-        }
-    }
-}
-
-impl std::error::Error for StageFrameError {}
-
-pub fn validate_stage_control_request(
-    frame: &proto::stage::StageControlRequest,
-) -> Result<(), StageFrameError> {
-    validate_generation(frame.r#gen)?;
-    validate_endpoint_id(frame.requester_id.len())?;
-    if frame.command.is_none() {
-        return Err(StageFrameError::MissingStageControlCommand);
-    }
-    Ok(())
-}
-
-pub fn validate_stage_control_response(
-    frame: &proto::stage::StageControlResponse,
-) -> Result<(), StageFrameError> {
-    validate_generation(frame.r#gen)?;
-    if frame.response.is_none() {
-        return Err(StageFrameError::MissingStageControlResponse);
-    }
-    Ok(())
-}
-
-pub fn validate_stage_transport_open(
-    frame: &proto::stage::StageTransportOpen,
-) -> Result<(), StageFrameError> {
-    validate_generation(frame.r#gen)?;
-    validate_endpoint_id(frame.requester_id.len())?;
-    if frame.topology_id.is_empty() || frame.run_id.is_empty() || frame.stage_id.is_empty() {
-        return Err(StageFrameError::MissingStageTransportTarget);
-    }
-    Ok(())
-}
-
-pub fn validate_stage_artifact_transfer_request(
-    frame: &proto::stage::StageArtifactTransferRequest,
-) -> Result<(), StageFrameError> {
-    validate_generation(frame.r#gen)?;
-    validate_endpoint_id(frame.requester_id.len())?;
-    if frame.topology_id.is_empty()
-        || frame.run_id.is_empty()
-        || frame.stage_id.is_empty()
-        || !frame.package_ref.starts_with("hf://")
-    {
-        return Err(StageFrameError::MissingStageArtifactTarget);
-    }
-    validate_artifact_digest(&frame.manifest_sha256)?;
-    if let Some(expected_sha) = frame.expected_sha256.as_deref() {
-        validate_artifact_digest(expected_sha)?;
-    }
-    if frame.expected_size.is_some_and(|size| frame.offset > size) {
-        return Err(StageFrameError::InvalidArtifactOffset);
-    }
-    validate_safe_relative_artifact_path(&frame.relative_path)?;
-    Ok(())
-}
-
-pub fn validate_stage_artifact_transfer_response(
-    frame: &proto::stage::StageArtifactTransferResponse,
-) -> Result<(), StageFrameError> {
-    validate_generation(frame.r#gen)?;
-    if let Some(sha256) = frame.sha256.as_deref() {
-        validate_artifact_digest(sha256)?;
-    }
-    Ok(())
-}
-
-fn validate_generation(r#gen: u32) -> Result<(), StageFrameError> {
-    if r#gen != STAGE_PROTOCOL_GENERATION {
-        return Err(StageFrameError::BadGeneration { got: r#gen });
-    }
-    Ok(())
-}
-
-fn validate_artifact_digest(value: &str) -> Result<(), StageFrameError> {
-    if value.len() != 64 || !value.chars().all(|ch| ch.is_ascii_hexdigit()) {
-        return Err(StageFrameError::InvalidArtifactDigestLength { got: value.len() });
-    }
-    Ok(())
-}
-
-fn validate_safe_relative_artifact_path(path: &str) -> Result<(), StageFrameError> {
-    use std::path::{Component, Path};
-
-    if path.trim().is_empty() {
-        return Err(StageFrameError::InvalidArtifactPath);
-    }
-    let path = Path::new(path);
-    let mut components = path.components();
-    let Some(first) = components.next() else {
-        return Err(StageFrameError::InvalidArtifactPath);
-    };
-    if !matches!(first, Component::Normal(_))
-        || !components.all(|component| matches!(component, Component::Normal(_)))
-    {
-        return Err(StageFrameError::InvalidArtifactPath);
-    }
-    Ok(())
-}
-
-fn validate_endpoint_id(len: usize) -> Result<(), StageFrameError> {
-    if len != 32 {
-        return Err(StageFrameError::InvalidEndpointId { got: len });
-    }
-    Ok(())
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum MessageKind {
-    Ready,
-    PrefillChunk,
-    FinalPrefillChunk,
-    DecodeToken,
-    StateImport,
-    StateExport,
-    Ack,
-    TokenReply,
-    Stop,
-    Error,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StageIdentity {
-    pub run_id: String,
-    pub request_id: String,
-    pub session_id: String,
-    pub topology_id: String,
-    pub stage_id: String,
-    pub stage_index: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum LoadMode {
-    RuntimeSlice,
-    LayerPackage,
-    ArtifactSlice,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum FlashAttentionType {
-    #[default]
-    Auto,
-    Disabled,
-    Enabled,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StageConfig {
-    pub run_id: String,
-    pub topology_id: String,
-    pub model_id: String,
-    #[serde(default)]
-    pub package_ref: Option<String>,
-    #[serde(default)]
-    pub manifest_sha256: Option<String>,
-    #[serde(default)]
-    pub source_model_path: Option<String>,
-    #[serde(default)]
-    pub source_model_sha256: Option<String>,
-    #[serde(default)]
-    pub source_model_bytes: Option<u64>,
-    #[serde(default)]
-    pub materialized_path: Option<String>,
-    #[serde(default)]
-    pub materialized_pinned: bool,
-    #[serde(default)]
-    pub model_path: Option<String>,
-    #[serde(default)]
-    pub projector_path: Option<String>,
-    pub stage_id: String,
-    pub stage_index: u32,
-    pub layer_start: u32,
-    pub layer_end: u32,
-    #[serde(default = "default_ctx_size")]
-    pub ctx_size: u32,
-    #[serde(default = "default_lane_count")]
-    pub lane_count: u32,
-    #[serde(default)]
-    pub n_batch: Option<u32>,
-    #[serde(default)]
-    pub n_ubatch: Option<u32>,
-    #[serde(default)]
-    pub n_gpu_layers: i32,
-    #[serde(default)]
-    pub mmap: Option<bool>,
-    #[serde(default)]
-    pub mlock: bool,
-    #[serde(default = "default_cache_type")]
-    pub cache_type_k: String,
-    #[serde(default = "default_cache_type")]
-    pub cache_type_v: String,
-    #[serde(default)]
-    pub flash_attn_type: FlashAttentionType,
-    #[serde(default)]
-    pub filter_tensors_on_load: bool,
-    #[serde(default)]
-    pub selected_device: Option<StageDevice>,
-    #[serde(default)]
-    pub kv_cache: Option<StageKvCacheConfig>,
-    #[serde(default = "default_native_mtp_enabled")]
-    pub native_mtp_enabled: bool,
-    pub load_mode: LoadMode,
-    pub bind_addr: String,
-    #[serde(default)]
-    pub upstream: Option<PeerConfig>,
-    #[serde(default)]
-    pub downstream: Option<PeerConfig>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StageDevice {
-    pub backend_device: String,
-    #[serde(default)]
-    pub stable_id: Option<String>,
-    #[serde(default)]
-    pub index: Option<usize>,
-    #[serde(default)]
-    pub vram_bytes: Option<u64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum StageKvCacheMode {
-    Disabled,
-    Auto,
-    Record,
-    LookupRecord,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum StageKvCachePayload {
-    Auto,
-    ResidentKv,
-    KvRecurrent,
-    FullState,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StageKvCacheConfig {
-    #[serde(default = "default_kv_cache_mode")]
-    pub mode: StageKvCacheMode,
-    #[serde(default = "default_kv_cache_payload")]
-    pub payload: StageKvCachePayload,
-    #[serde(default = "default_kv_cache_max_entries")]
-    pub max_entries: usize,
-    #[serde(default)]
-    pub max_bytes: u64,
-    #[serde(default = "default_kv_cache_min_tokens")]
-    pub min_tokens: u64,
-    #[serde(default = "default_kv_cache_shared_stride_tokens")]
-    pub shared_prefix_stride_tokens: u64,
-    #[serde(default = "default_kv_cache_shared_record_limit")]
-    pub shared_prefix_record_limit: u64,
-}
-
-fn default_kv_cache_mode() -> StageKvCacheMode {
-    StageKvCacheMode::Auto
-}
-
-fn default_kv_cache_payload() -> StageKvCachePayload {
-    StageKvCachePayload::Auto
-}
-
-fn default_kv_cache_max_entries() -> usize {
-    64
-}
-
-fn default_kv_cache_min_tokens() -> u64 {
-    64
-}
-
-fn default_kv_cache_shared_stride_tokens() -> u64 {
-    128
-}
-
-fn default_kv_cache_shared_record_limit() -> u64 {
-    2
-}
-
-fn default_ctx_size() -> u32 {
-    512
-}
-
-fn default_lane_count() -> u32 {
-    4
-}
-
-fn default_cache_type() -> String {
-    "f16".to_string()
-}
-
-fn default_native_mtp_enabled() -> bool {
-    true
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct PeerConfig {
-    pub stage_id: String,
-    pub stage_index: u32,
-    pub endpoint: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StageTopology {
-    pub topology_id: String,
-    pub model_id: String,
-    pub stages: Vec<StageTopologyEntry>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StageTopologyEntry {
-    pub stage_id: String,
-    pub stage_index: u32,
-    pub host: Option<String>,
-    pub endpoint: String,
-    pub layer_start: u32,
-    pub layer_end: u32,
-    pub load_mode: LoadMode,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ActivationDType {
-    Unknown,
-    F32,
-    F16,
-    Bf16,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ActivationLayout {
-    Opaque,
-    TokenMajor,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct ActivationDescriptor {
-    pub version: u32,
-    pub dtype: ActivationDType,
-    pub layout: ActivationLayout,
-    pub producer_stage_index: i32,
-    pub layer_start: i32,
-    pub layer_end: i32,
-    pub token_count: u32,
-    pub sequence_count: u32,
-    pub payload_bytes: u64,
-    #[serde(default)]
-    pub flags: u64,
-    #[serde(default)]
-    pub payload_sha256: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(tag = "message_type", rename_all = "snake_case")]
-pub enum StageMessage {
-    Ready(ReadyMessage),
-    PrefillChunk(PrefillChunkMessage),
-    FinalPrefillChunk(FinalPrefillChunkMessage),
-    DecodeToken(DecodeTokenMessage),
-    StateImport(StateImportMessage),
-    StateExport(StateExportMessage),
-    Ack(AckMessage),
-    TokenReply(TokenReplyMessage),
-    Stop(StopMessage),
-    Error(ErrorMessage),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct MessageBase {
-    pub schema_version: u32,
-    pub run_id: String,
-    pub request_id: String,
-    pub session_id: String,
-    pub stage_id: String,
-    pub stage_index: u32,
-    pub topology_id: String,
-    #[serde(default)]
-    pub model_id: Option<String>,
-    #[serde(default)]
-    pub tokenizer_id: Option<String>,
-    #[serde(default)]
-    pub chat_template_id: Option<String>,
-    #[serde(default)]
-    pub seq: Option<u64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct ReadyMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub layer_start: u32,
-    pub layer_end: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct PrefillChunkMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub token_ids: Vec<i32>,
-    pub prompt_token_start: u32,
-    #[serde(default)]
-    pub activation_dtype: Option<String>,
-    #[serde(default)]
-    pub activation_bytes: Option<u64>,
-    #[serde(default)]
-    pub activation: Option<ActivationDescriptor>,
-    #[serde(default)]
-    pub activation_ref: Option<String>,
-    #[serde(default)]
-    pub is_final: Option<bool>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct FinalPrefillChunkMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub token_ids: Vec<i32>,
-    pub prompt_token_start: u32,
-    pub is_final: bool,
-    #[serde(default)]
-    pub activation_dtype: Option<String>,
-    #[serde(default)]
-    pub activation_bytes: Option<u64>,
-    #[serde(default)]
-    pub activation: Option<ActivationDescriptor>,
-    #[serde(default)]
-    pub activation_ref: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct DecodeTokenMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub token_id: i32,
-    pub decode_index: u32,
-    #[serde(default)]
-    pub activation_dtype: Option<String>,
-    #[serde(default)]
-    pub activation_bytes: Option<u64>,
-    #[serde(default)]
-    pub activation: Option<ActivationDescriptor>,
-    #[serde(default)]
-    pub activation_ref: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StateImportMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub layer_start: u32,
-    pub layer_end: u32,
-    pub state_bytes: u64,
-    #[serde(default)]
-    pub state_sha256: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StateExportMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub layer_start: u32,
-    pub layer_end: u32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct AckMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub acked_seq: u64,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct TokenReplyMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub token_id: i32,
-    #[serde(default)]
-    pub decode_index: Option<u32>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct StopMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-pub struct ErrorMessage {
-    #[serde(flatten)]
-    pub base: MessageBase,
-    pub error_code: String,
-    pub error_message: String,
-}
-
-impl StageConfig {
-    pub fn ready_message(&self) -> StageMessage {
-        StageMessage::Ready(ReadyMessage {
-            base: MessageBase {
-                schema_version: SCHEMA_VERSION,
-                run_id: self.run_id.clone(),
-                request_id: "stage-ready".to_string(),
-                session_id: "stage-lifecycle".to_string(),
-                stage_id: self.stage_id.clone(),
-                stage_index: self.stage_index,
-                topology_id: self.topology_id.clone(),
-                model_id: Some(self.model_id.clone()),
-                tokenizer_id: None,
-                chat_template_id: None,
-                seq: Some(0),
-            },
-            layer_start: self.layer_start,
-            layer_end: self.layer_end,
-        })
-    }
-}
-
-impl StageMessage {
-    pub fn base(&self) -> &MessageBase {
-        match self {
-            Self::Ready(message) => &message.base,
-            Self::PrefillChunk(message) => &message.base,
-            Self::FinalPrefillChunk(message) => &message.base,
-            Self::DecodeToken(message) => &message.base,
-            Self::StateImport(message) => &message.base,
-            Self::StateExport(message) => &message.base,
-            Self::Ack(message) => &message.base,
-            Self::TokenReply(message) => &message.base,
-            Self::Stop(message) => &message.base,
-            Self::Error(message) => &message.base,
-        }
-    }
-
-    pub fn kind(&self) -> MessageKind {
-        match self {
-            Self::Ready(_) => MessageKind::Ready,
-            Self::PrefillChunk(_) => MessageKind::PrefillChunk,
-            Self::FinalPrefillChunk(_) => MessageKind::FinalPrefillChunk,
-            Self::DecodeToken(_) => MessageKind::DecodeToken,
-            Self::StateImport(_) => MessageKind::StateImport,
-            Self::StateExport(_) => MessageKind::StateExport,
-            Self::Ack(_) => MessageKind::Ack,
-            Self::TokenReply(_) => MessageKind::TokenReply,
-            Self::Stop(_) => MessageKind::Stop,
-            Self::Error(_) => MessageKind::Error,
-        }
-    }
-
-    pub fn ack_for(&self, stage: &StageConfig) -> StageMessage {
-        let base = self.base();
-        StageMessage::Ack(AckMessage {
-            base: MessageBase {
-                schema_version: SCHEMA_VERSION,
-                run_id: base.run_id.clone(),
-                request_id: base.request_id.clone(),
-                session_id: base.session_id.clone(),
-                stage_id: stage.stage_id.clone(),
-                stage_index: stage.stage_index,
-                topology_id: stage.topology_id.clone(),
-                model_id: Some(stage.model_id.clone()),
-                tokenizer_id: base.tokenizer_id.clone(),
-                chat_template_id: base.chat_template_id.clone(),
-                seq: base.seq,
-            },
-            acked_seq: base.seq.unwrap_or(0),
-        })
-    }
-}
+pub use config::{
+    ActivationDType, ActivationDescriptor, ActivationLayout, FlashAttentionType, LoadMode,
+    PeerConfig, StageConfig, StageDevice, StageIdentity, StageKvCacheConfig, StageKvCacheMode,
+    StageKvCachePayload, StageTopology, StageTopologyEntry,
+};
+pub use messages::{
+    AckMessage, DecodeTokenMessage, ErrorMessage, FinalPrefillChunkMessage, MessageBase,
+    MessageKind, PrefillChunkMessage, ReadyMessage, StageMessage, StateExportMessage,
+    StateImportMessage, StopMessage, TokenReplyMessage,
+};
+pub use validation::{
+    MAX_STAGE_FRAME_BYTES, MAX_VERIFY_WINDOW_PIPELINE_DEPTH, SCHEMA_VERSION, STAGE_ALPN_V2,
+    STAGE_PROTOCOL_GENERATION, STAGE_STREAM_ARTIFACT_TRANSFER, STAGE_STREAM_CONTROL,
+    STAGE_STREAM_TRANSPORT, STAGE_SUBPROTOCOL_FEATURE_ARTIFACT_TRANSFER,
+    STAGE_SUBPROTOCOL_FEATURE_STAGE_CONTROL, STAGE_SUBPROTOCOL_FEATURE_STAGE_GENERATION,
+    STAGE_SUBPROTOCOL_FEATURE_STAGE_PROTOCOL_GENERATION_V4, STAGE_SUBPROTOCOL_FEATURE_STATUS_LIST,
+    STAGE_SUBPROTOCOL_MAJOR, STAGE_SUBPROTOCOL_NAME, StageFrameError,
+    validate_stage_artifact_transfer_request, validate_stage_artifact_transfer_response,
+    validate_stage_control_request, validate_stage_control_response, validate_stage_transport_open,
+};
 
 #[cfg(test)]
 mod tests {
